refactor: core:domain 모듈 전환 및 패키지 리팩토링
- core:domain 모듈을 Android Application에서 Java 라이브러리(java-library)로 변경
- com.example.domain 패키지 경로를 com.afternote.core.domain으로 수정
- 매니페스트, 테마, 아이콘 등 불필요한 Android 리소스 파일 삭제
- PhotoUploadRepository, MemorialVideoUploadRepository 등 미디어 업로드 인터페이스 추가
- feature:afternote:domain 모듈에 core:domain 의존성 추가 및 임포트 경로 수정
- 신규 모듈 lib 및 domain 프로젝트 설정 추가
