+++ core/domain/src/main/java/com/afternote/core/domain/usecase/UserSessionRepository.kt
@@ -1,0 +1,5 @@
+package com.afternote.core.domain.usecase
+
+fun interface UserSessionRepository {
+    fun getUserId(): Long?
+}
