--- core/domain/src/main/java/com/example/domain/repository/UserRepository.kt
@@ -1,95 +1,0 @@
-package com.example.domain.repository
-
-import com.example.domain.model.DeliveryCondition
-import com.example.domain.model.DeliveryConditionType
-import com.example.domain.model.PushSettings
-import com.example.domain.model.ReceiverDailyQuestionsResult
-import com.example.domain.model.ReceiverDetail
-import com.example.domain.model.ReceiverListItem
-import com.example.domain.model.ReceiverMindRecordsResult
-import com.example.domain.model.UserProfile
-
-/**
- * User 도메인 Repository 인터페이스. (스웨거 기준)
- */
-interface UserRepository {
-    suspend fun getMyProfile(userId: Long): Result<UserProfile>
-
-    suspend fun updateMyProfile(
-        userId: Long,
-        name: String?,
-        phone: String?,
-        profileImageUrl: String?,
-    ): Result<UserProfile>
-
-    /**
-     * DELETE /users/me — 회원 탈퇴. 로그인한 사용자의 계정을 삭제합니다. 모든 데이터가 영구 삭제되며 복구할 수 없습니다.
-     */
-    suspend fun withdrawAccount(): Result<Unit>
-
-    /** GET /users/push-settings — 푸시 알림 설정 조회. */
-    suspend fun getMyPushSettings(userId: Long): Result<PushSettings>
-
-    suspend fun updateMyPushSettings(
-        userId: Long,
-        timeLetter: Boolean?,
-        mindRecord: Boolean?,
-        afterNote: Boolean?,
-    ): Result<PushSettings>
-
-    /** GET /users/receivers — 수신인 목록 조회. 로그인한 사용자가 등록한 수신인 목록을 조회합니다. */
-    suspend fun getReceivers(userId: Long): Result<List<ReceiverListItem>>
-
-    suspend fun registerReceiver(
-        name: String,
-        relation: String,
-        phone: String?,
-        email: String?,
-    ): Result<Long>
-
-    suspend fun getReceiverDetail(receiverId: Long): Result<ReceiverDetail>
-
-    suspend fun updateReceiver(
-        receiverId: Long,
-        name: String,
-        relation: String,
-        phone: String?,
-        email: String?,
-    ): Result<Unit>
-
-    suspend fun getReceiverDailyQuestions(
-        receiverId: Long,
-        page: Int,
-        size: Int,
-    ): Result<ReceiverDailyQuestionsResult>
-
-    /**
-     * 수신인별 마음의 기록 전체 조회 (일기, 깊은 생각, 데일리 질문 답변).
-     * GET /users/receivers/{receiverId}/mind-records
-     */
-    suspend fun getReceiverMindRecords(
-        receiverId: Long,
-        page: Int,
-        size: Int,
-    ): Result<ReceiverMindRecordsResult>
-
-    /**
-     * GET /users/delivery-condition — 로그인한 사용자의 전달 조건 설정 조회.
-     */
-    suspend fun getDeliveryCondition(): Result<DeliveryCondition>
-
-    /**
-     * PATCH /users/delivery-condition — 로그인한 사용자의 전달 조건 설정/변경.
-     *
-     * @param conditionType 전달 조건 타입
-     * @param inactivityPeriodDays 비활동 기간(일), INACTIVITY일 때 사용
-     * @param specificDate 특정 날짜(yyyy-MM-dd), SPECIFIC_DATE일 때 사용
-     * @param leaveMessage 마지막 인사말 (수신자에게 전달되는 메시지)
-     */
-    suspend fun updateDeliveryCondition(
-        conditionType: DeliveryConditionType,
-        inactivityPeriodDays: Int?,
-        specificDate: String?,
-        leaveMessage: String? = null,
-    ): Result<DeliveryCondition>
-}
