--- core/domain/src/main/java/com/example/domain/usecase/UserSessionRepository.kt
@@ -1,5 +1,0 @@
-package com.example.domain.usecase
-
-fun interface UserSessionRepository {
-    fun getUserId(): Long?
-}
