+++ domain/src/main/java/com/afternote/core/domain/MyClass.kt
@@ -1,0 +1,3 @@
+package com.afternote.core.domain
+
+class MyClass
