+++ feature/afternote/domain/build.gradle.kts
@@ -13,4 +13,5 @@ kotlin {
 }
 dependencies {
     implementation(libs.javax.inject)
+    implementation(project(":core:domain"))
 }
