+++ feature/afternote/domain/src/main/java/com/afternote/feature/afternote/domain/usecase/UploadMemorialPhotoUseCase.kt
@@ -1,6 +1,6 @@
-package com.kuit.afternote.feature.afternote.domain.usecase
+package com.afternote.feature.afternote.domain.usecase
 
-import com.kuit.afternote.domain.repository.PhotoUploadRepository
+import com.afternote.core.domain.repository.PhotoUploadRepository
 import javax.inject.Inject
 
 private const val DIRECTORY_AFTERNOTES = "afternotes"
