+++ feature/afternote/domain/src/main/java/com/afternote/feature/afternote/domain/usecase/UploadMemorialThumbnailUseCase.kt
@@ -1,6 +1,6 @@
-package com.kuit.afternote.feature.afternote.domain.usecase
+package com.afternote.feature.afternote.domain.usecase
 
-import com.kuit.afternote.domain.repository.MemorialThumbnailUploadRepository
+import com.afternote.feature.afternote.domain.repository.MemorialThumbnailUploadRepository
 import javax.inject.Inject
 
 /**
