+++ lib/build.gradle.kts
@@ -11,6 +11,3 @@ kotlin {
         jvmTarget = org.jetbrains.kotlin.gradle.dsl.JvmTarget.JVM_11
     }
 }
-dependencies {
-    implementation(libs.javax.inject)
-}
