+++ lib/src/main/java/com/example/lib/MyClass.kt
@@ -1,0 +1,3 @@
+package com.example.lib
+
+class MyClass
