+++ settings.gradle.kts
@@ -45,4 +45,5 @@ include(":feature:onboarding:domain")
 include(":core:datastore")
 include(":core:model")
 include(":core:common")
+include(":lib")
 include(":core:domain")
